requirements.txt
asgiref==3.2.7
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
dj-database-url==0.5.0
Django==3.0.6
django-cors-headers==3.3.0
django-extensions==2.2.9
djangorestframework==3.11.0
drf-jwt==1.16.1
gunicorn==20.0.4
idna==2.9
psycopg2-binary==2.8.5
pycparser==2.20
pydot==1.4.1
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2020.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.1.0